Finally got the EM working but there is a problem with precision. C cannot give the precision we need with the exp function. May have to explore alternative options
